+++ src/codemagic/tools/firebase/firebase_action.py
@@ -3,9 +3,8 @@ from abc import ABCMeta
 from abc import abstractmethod
 
 from codemagic.firebase.api_client import FirebaseApiClient
-
-from ..arguments import AppId
-from ..arguments import ProjectId
+from codemagic.firebase.resource_managers.release_manager import AppId
+from codemagic.firebase.resource_managers.release_manager import ProjectId
 
 
 class FirebaseAction(metaclass=ABCMeta):
@@ -26,6 +25,6 @@ class FirebaseAction(metaclass=ABCMeta):
         project_id: ProjectId,
         app_id: AppId,
     ) -> int:
-        from ..actions import GetLatestBuildVersionAction
+        from .actions import GetLatestBuildVersionAction
         _ = GetLatestBuildVersionAction.get_latest_build_version  # Implementation
         raise NotImplementedError()