+++ tests/firebase/test_api_client.py
@@ -11,7 +11,9 @@ from unittest.mock import patch
 import pytest
 
 from codemagic.firebase import FirebaseApiClient
+from codemagic.firebase.resource_managers.release_manager import AppId
 from codemagic.firebase.resource_managers.release_manager import FirebaseReleaseManager
+from codemagic.firebase.resource_managers.release_manager import ProjectId
 from codemagic.firebase.resource_managers.release_manager import ReleaseParentIdentifier
 from codemagic.firebase.resources import Release
 from codemagic.firebase.resources import ReleaseNotes
@@ -46,13 +48,18 @@ def mock_release():
 
 
 @pytest.fixture
-def project_id():
-    return '228333310124'
+def app_id():
+    return AppId('1:228333310124:ios:5e439e0d0231a788ac8f09')
 
 
 @pytest.fixture
-def app_id():
-    return '1:228333310124:ios:5e439e0d0231a788ac8f09'
+def project_id():
+    return ProjectId('228333310124')
+
+
+@pytest.fixture
+def release_parent_identifier(project_id, app_id):
+    return ReleaseParentIdentifier(project_id, app_id)
 
 
 @pytest.fixture
@@ -81,19 +88,17 @@ def test_release(mock_release_response_data, mock_release):
 
 
 @pytest.mark.skipif(not os.environ.get('RUN_LIVE_API_TESTS'), reason='Live Firebase API access')
-def test_list_releases_live(project_id, app_id, credentials, client, mock_release):
+def test_list_releases_live(release_parent_identifier, credentials, client, mock_release):
     order_by = FirebaseReleaseManager.OrderBy.create_time_asc
-    parent_identifier = ReleaseParentIdentifier(project_id, app_id)
-    releases = client.releases.list(parent_identifier, order_by=order_by, page_size=2)
+    releases = client.releases.list(release_parent_identifier, order_by=order_by, page_size=2)
 
     assert releases[0] == mock_release
     assert releases[1].buildVersion == 71
 
 
 @pytest.mark.skipif(not os.environ.get('RUN_LIVE_API_TESTS'), reason='Live Firebase API access')
-def test_list_releases_pagination_live(project_id, app_id, credentials, client, mock_release):
-    parent_identifier = ReleaseParentIdentifier(project_id, app_id)
-    releases = client.releases.list(parent_identifier, page_size=1)
+def test_list_releases_pagination_live(release_parent_identifier, credentials, client, mock_release):
+    releases = client.releases.list(release_parent_identifier, page_size=1)
 
     assert releases[0].buildVersion == 71
     assert releases[1] == mock_release
@@ -131,17 +136,15 @@ def mock_releases_api_resource(mock_release_response_data):
         yield mock_resource
 
 
-def test_list_releases(project_id, app_id, mock_client, mock_release, mock_releases_api_resource):
+def test_list_releases(release_parent_identifier, mock_client, mock_release, mock_releases_api_resource):
     order_by = FirebaseReleaseManager.OrderBy.create_time_asc
-    parent_identifier = ReleaseParentIdentifier(project_id, app_id)
-    releases = mock_client.releases.list(parent_identifier, order_by=order_by, page_size=2)
+    releases = mock_client.releases.list(release_parent_identifier, order_by=order_by, page_size=2)
 
     assert releases[0] == mock_release
     assert releases[1].buildVersion == 71
 
 
-def test_list_releases_limit(project_id, app_id, mock_client, mock_release, mock_releases_api_resource):
-    parent_identifier = ReleaseParentIdentifier(project_id, app_id)
-    releases = mock_client.releases.list(parent_identifier, limit=1)
+def test_list_releases_limit(release_parent_identifier, mock_client, mock_release, mock_releases_api_resource):
+    releases = mock_client.releases.list(release_parent_identifier, limit=1)
 
     assert len(releases) == 1